Create and implement PlayerAPI (#13)
* Create Player class

* Implement validation for player name

To create player any name must be given. For this purpose validation had
been implemented. Player name can not be null.

* Update player name validation

* Implement sign validation for null case

* Update sign validation for default case

* Add field points

* Implement equals()

* Implement hashCode()

* Add constructor with parameter points

There can be possibilty that we want to creat player with given number
of points. More to that we have to be sure that player have at least 0
or more points.

* Create setPlayerPoints() method

* Override toString()

* Create PlayerFactory

* Fix test cases

* Move classes to proper packages

* Reduce visibility of PlayerFactory

* Create and implement PlayerAPI

* Test PlayerAPIImpl